Add version utilizing macro by Piepho
Popular way to present the result of mean comparisons is by attaching super-scripted letters to the means, with a common letter on two means indicating that they are not significantly different. When the standard error of a difference is not constant, the algorithm traditionally used for generating such displays may fail to represent all significant differences. This macro is based on a so-called insert-and-absorb algorithm that is guaranteed to always truthfully represent all significant differences.